backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
notion-client==2.2.1
PyGithub==2.5.0
python-dotenv==1.0.1
httpx==0.27.0
